Add FormattableLedgerCommand
- abstracts the ledger parsing completeley from the commands
- helps with testing
