+++ Sources/SwiftBeanCountCLILibrary/SubCommands/Stats.swift
@@ -1,9 +1,7 @@
 import ArgumentParser
-import Foundation
-import Rainbow
 import SwiftBeanCountModel
 
-struct Stats: LedgerCommand, FormattableCommand {
+struct Stats: FormattableLedgerCommand {
 
     static var configuration = CommandConfiguration(abstract: "Statistics of a ledger (e.g. # of transactions)")
 
@@ -11,12 +9,7 @@ struct Stats: LedgerCommand, FormattableCommand {
     @ArgumentParser.Option(name: [.short, .long], default: .table, help: "Output format. \(Self.supportedFormats())") var format: Format
     @ArgumentParser.Flag(help: Self.noColorHelp()) var noColor: Bool
 
-    func getResult() throws -> FormattableResult {
-        let start = Date.timeIntervalSinceReferenceDate
-        let ledger = try parseLedger()
-        let end = Date.timeIntervalSinceReferenceDate
-        let time = end - start
-
+    func getResult(from ledger: Ledger, parsingDuration: Double) -> FormattableResult {
         let values: [[String]] = [
             ["Transactions", String(ledger.transactions.count)],
             ["Accounts", String(ledger.accounts.count)],
@@ -32,11 +25,7 @@ struct Stats: LedgerCommand, FormattableCommand {
             ["Plugins", String(ledger.plugins.count)]
         ]
 
-        var footer: String?
-        if format != .csv {
-            footer = String(format: "Parsing time: %.3f sec", time)
-        }
-
+        let footer = String(format: "Parsing time: %.3f sec", parsingDuration)
         return FormattableResult(title: "Statistics", columns: ["Type", "Number"], values: values, footer: footer)
     }
